gui/app/pyproject.toml
[project]
name = "pypodpro"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jellyfin-apiclient-python>=1.11.0",
    "pygame-ce>=2.5.5",
]
